tools/pest_detection/requirements.txt
# Pest Detection Tool Requirements

# Deep Learning Framework
# Updated to address CVE vulnerabilities in torch < 2.6.0
torch>=2.6.0
torchvision>=0.19.0

# YOLO Model
# Updated to latest compatible version
ultralytics>=8.3.0

# Image Processing
opencv-python>=4.10.0
pillow>=10.4.0

# Beam Cloud (for deployment)
# Install separately: pip install beam-client

# Testing Dependencies
requests>=2.32.0
